Fix crash on shutdown if no DLQ file has been opened (#33)
This commit fixed an  NullPointerException happening in DeadLetterQueueReader::getCurrentSegment or DeadLetterQueueReader::getCurrentPosition when nothing has been read yet. This causes the entirety of Logstash to crash and then, aggregate maps may not be saved.

This commit works by introducing a simple boolean, readerHasState, which is set to true when an event is received (and when state is explicitly set). If it is false, then the plugin will not attempt to access the state on shutdown.